feat(EMS-3631): various declaration constant alignments
