+++ src/ui/server/constants/declarations/index.test.ts
@@ -1,0 +1,14 @@
+import DECLARATIONS from '.';
+import VERSIONS from './versions';
+import getLatestDeclarationVersion from './get-latest-declarations';
+
+describe('server/constants/declarations/', () => {
+  it('should return an object', () => {
+    const expected = {
+      VERSIONS,
+      LATEST_DECLARATIONS: getLatestDeclarationVersion(),
+    };
+
+    expect(DECLARATIONS).toEqual(expected);
+  });
+});
